requirements.txt
joblib==1.2.0
matplotlib==3.6.2
opencv_python==4.6.0.66
Pillow==9.5.0
